+++ client/cody/src/completions/context.ts
@@ -1,10 +1,9 @@
-import path from 'path'
-
 import * as vscode from 'vscode'
 
 import { CodebaseContext } from '@sourcegraph/cody-shared/src/codebase-context'
 
-import { JaccardMatch, bestJaccardMatch } from './bestJaccardMatch'
+import { getContextFromEmbeddings } from './context-embeddings'
+import { getContextFromCurrentEditor } from './context-local'
 import { History } from './history'
 
 /**
@@ -15,115 +14,47 @@ export interface ReferenceSnippet {
     content: string
 }
 
-interface JaccardMatchWithFilename extends JaccardMatch {
-    fileName: string
-}
-
 interface GetContextOptions {
     currentEditor: vscode.TextEditor
     history: History
-    targetText: string
-    windowSize: number
+    prefix: string
+    suffix: string
+    jaccardDistanceWindowSize: number
     maxChars: number
     codebaseContext: CodebaseContext
 }
 
 export async function getContext(options: GetContextOptions): Promise<ReferenceSnippet[]> {
-    // TODO(valery): use codebaseContext here.
-    const { currentEditor, history, targetText, windowSize, maxChars } = options
-    const files = await getRelevantFiles(currentEditor, history)
+    const { maxChars } = options
 
-    const matches: JaccardMatchWithFilename[] = []
-    for (const { uri, contents } of files) {
-        const match = bestJaccardMatch(targetText, contents, windowSize)
-        if (!match) {
-            continue
-        }
+    /**
+     * The embeddings context is sync to retrieve to keep the completions latency minumal.
+     * If it's not available in cache yet, we'll retrieve it in the background and cache it for future use.
+     */
+    const embeddingsMatches = getContextFromEmbeddings(options)
+    const editorMatches = await getContextFromCurrentEditor(options)
 
-        matches.push({
-            // Use relative path to remove redundant information from the prompts and
-            // keep in sync with embeddings search resutls which use relatve to repo root paths.
-            fileName: path.normalize(vscode.workspace.asRelativePath(uri.fsPath)),
-            ...match,
-        })
-    }
-
-    matches.sort((a, b) => b.score - a.score)
-
+    const usedFilenames = new Set<string>()
     const context: ReferenceSnippet[] = []
     let totalChars = 0
-    for (const match of matches) {
-        if (totalChars + match.content.length > maxChars) {
-            break
+
+    /**
+     * Iterate over matches and add them to the context.
+     * Discard editor matches for files with embedding matches.
+     */
+    for (const match of [...embeddingsMatches, ...editorMatches]) {
+        const existingMatch = usedFilenames.has(match.fileName)
+
+        if (!existingMatch) {
+            usedFilenames.add(match.fileName)
+
+            if (totalChars + match.content.length > maxChars) {
+                break
+            }
+            context.push(match)
+            totalChars += match.content.length
         }
-        context.push(match)
-        totalChars += match.content.length
     }
 
     return context
-}
-
-interface FileContents {
-    uri: vscode.Uri
-    contents: string
-}
-
-/**
- * Loads all relevant files for for a given text editor. Relevant files are defined as:
- *
- * - All currently open tabs matching the same language
- * - The last 10 files that were edited matching the same language
- *
- * For every file, we will load up to 10.000 lines to avoid OOMing when working with very large
- * files.
- */
-async function getRelevantFiles(currentEditor: vscode.TextEditor, history: History): Promise<FileContents[]> {
-    const files: FileContents[] = []
-
-    const curLang = currentEditor.document.languageId
-    if (!curLang) {
-        return []
-    }
-
-    function addDocument(document: vscode.TextDocument): void {
-        if (document.uri === currentEditor.document.uri) {
-            // omit current file
-            return
-        }
-        if (document.languageId !== curLang) {
-            // TODO(beyang): handle JavaScript <-> TypeScript and verify this works for C header files
-            // omit files of other languages
-            return
-        }
-
-        // TODO(philipp-spiess): Find out if we have a better approach to truncate very large files.
-        const endLine = Math.min(document.lineCount, 10_000)
-        const range = new vscode.Range(0, 0, endLine, 0)
-
-        files.push({
-            uri: document.uri,
-            contents: document.getText(range),
-        })
-    }
-
-    const documents = vscode.workspace.textDocuments
-    for (const document of documents) {
-        if (document.fileName.endsWith('.git')) {
-            // The VS Code API returns fils with the .git suffix for every open file
-            continue
-        }
-        addDocument(document)
-    }
-
-    await Promise.all(
-        history.lastN(10, curLang, [currentEditor.document.uri, ...files.map(f => f.uri)]).map(async item => {
-            try {
-                const document = await vscode.workspace.openTextDocument(item.document.uri)
-                addDocument(document)
-            } catch (error) {
-                console.error(error)
-            }
-        })
-    )
-    return files
 }