+++ enterprise/cmd/cody-gateway/internal/httpapi/completions/openai.go
@@ -12,6 +12,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/cody-gateway/internal/limiter"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/cody-gateway/internal/notify"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/completions/client/openai"
+	"github.com/sourcegraph/sourcegraph/internal/conf/conftypes"
 )
 
 const openAIURL = "https://api.openai.com/v1/chat/completions"
@@ -30,7 +31,7 @@ func NewOpenAIHandler(
 		eventLogger,
 		rs,
 		rateLimitNotifier,
-		openai.ProviderName,
+		string(conftypes.CompletionsProviderNameOpenAI),
 		openAIURL,
 		allowedModels,
 		upstreamHandlerMethods[openaiRequest]{