+++ enterprise/dev/ci/push_all.sh
@@ -8,7 +8,7 @@ function preview_tags() {
 
   for tag in "${tags[@]}"; do
     for registry in "${registries[@]}"; do
-      echo -e "\t ${registry}/\$IMAGE:${qa_prefix}-${tag}"
+      echo -e "\t ${registry}/\$IMAGE:${tag}"
     done
   done
 }
@@ -44,8 +44,6 @@ prod_registries=(
 )
 
 date_fragment="$(date +%Y-%m-%d)"
-
-qa_prefix="bazel"
 
 dev_tags=(
   "${BUILDKITE_COMMIT:0:12}"
@@ -87,12 +85,12 @@ echo "--- done"
 
 dev_tags_args=""
 for t in "${dev_tags[@]}"; do
-  dev_tags_args="$dev_tags_args --tag ${qa_prefix}-${t}"
+  dev_tags_args="$dev_tags_args --tag ${t}"
 done
 prod_tags_args=""
 if $push_prod; then
   for t in "${prod_tags[@]}"; do
-    prod_tags_args="$prod_tags_args --tag ${qa_prefix}-${t}"
+    prod_tags_args="$prod_tags_args --tag ${t}"
   done
 fi